Env vars oidc
